[WIP] Refactor: database & repos
