+++ app/src/main/java/com/softyorch/taskapp/data/Resource.kt
@@ -1,4 +1,4 @@
-package com.softyorch.taskapp.data.data
+package com.softyorch.taskapp.data
 
 sealed class Resource<T>(val data: T? = null, val message: String? = null) {
     class Success<T>(data: T) : Resource<T>(data = data)
