+++ app/src/main/java/com/softyorch/taskapp/data/database/tasks/Task.kt
@@ -1,4 +1,4 @@
-package com.softyorch.taskapp.domain.model
+package com.softyorch.taskapp.data.database.tasks
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
