+++ app/src/main/java/com/softyorch/taskapp/di/UserDataModule.kt
@@ -2,8 +2,8 @@ package com.softyorch.taskapp.di
 
 import android.content.Context
 import androidx.room.Room
-import com.softyorch.taskapp.data.data.userdata.UserDataBase
-import com.softyorch.taskapp.data.data.userdata.UserDataBaseDao
+import com.softyorch.taskapp.data.database.userdata.UserDataBase
+import com.softyorch.taskapp.data.database.userdata.UserDataBaseDao
 import com.softyorch.taskapp.utils.*
 import dagger.Module
 import dagger.Provides
