+++ app/src/main/java/com/softyorch/taskapp/ui/activities/MainActivityViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.softyorch.taskapp.data.data.Resource
-import com.softyorch.taskapp.domain.repository.DatastoreRepository
+import com.softyorch.taskapp.data.Resource
+import com.softyorch.taskapp.data.repository.DatastoreRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
