+++ app/src/main/java/com/softyorch/taskapp/ui/components/fabCustom/FABCustomViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.softyorch.taskapp.data.data.Resource
-import com.softyorch.taskapp.domain.model.Task
-import com.softyorch.taskapp.domain.repository.DatastoreRepository
-import com.softyorch.taskapp.domain.repository.TaskRepository
+import com.softyorch.taskapp.data.Resource
+import com.softyorch.taskapp.data.database.tasks.Task
+import com.softyorch.taskapp.data.repository.DatastoreRepository
+import com.softyorch.taskapp.data.repository.TaskRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
