+++ app/src/main/java/com/softyorch/taskapp/ui/components/topAppBarCustom/TopAppBarCustomViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.softyorch.taskapp.data.data.Resource
-import com.softyorch.taskapp.domain.model.UserData
-import com.softyorch.taskapp.domain.repository.DatastoreRepository
+import com.softyorch.taskapp.data.Resource
+import com.softyorch.taskapp.data.database.userdata.UserData
+import com.softyorch.taskapp.data.repository.DatastoreRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.flowOn
