+++ app/src/main/java/com/softyorch/taskapp/ui/screens/detail/DetailScreenViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.softyorch.taskapp.data.data.Resource
-import com.softyorch.taskapp.domain.model.Task
-import com.softyorch.taskapp.domain.repository.TaskRepository
+import com.softyorch.taskapp.data.Resource
+import com.softyorch.taskapp.data.database.tasks.Task
+import com.softyorch.taskapp.data.repository.TaskRepository
 import com.softyorch.taskapp.utils.emptyString
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
